Fix: Move upload function to util
